+++ web/src/pages/Settings/Streams/StreamModal/StreamModal.jsx
@@ -132,10 +132,25 @@ InternetRadioSearch.propTypes = {
     onChange: PropTypes.func.isRequired,
 };
 
+function validateInput(streamFields, streamTemplate) {
+    streamTemplate['fields'].filter((f) => f.required).map( (f) => {
+        if(!streamFields[f.name]) {
+            throw new Error(`Required field not set: ${f.name}`);
+        }
+    });
+}
+
+validateInput.propTypes = {
+    streamFields: PropTypes.object.isRequired,
+    streamTemplate: PropTypes.object.isRequired,
+}
+
 const StreamModal = ({ stream, onClose, apply, del }) => {
     const [streamFields, setStreamFields] = React.useState(
         JSON.parse(JSON.stringify(stream))
     ); // set streamFields to copy of stream
+
+    const [errorMessage, setErrorMessage] = React.useState("");
 
     const streamTemplate = StreamTemplates.filter(
         (t) => t.type === stream.type
@@ -150,11 +165,18 @@ const StreamModal = ({ stream, onClose, apply, del }) => {
                 onClose();
             }}
             onAccept={() => {
+                try {
+                    validateInput(streamFields, streamTemplate);
+                } catch (error) {
+                    setErrorMessage(error.toString());
+                    return;
+                }
                 apply(streamFields);
                 onClose();
             }}
         >
             <div>
+                <span>{ errorMessage }</span>
                 <TextField
                     name="Name"
                     desc={NAME_DESC}